refactor: rename recevier to receiver
