+++ oap-server/server-receiver-plugin/configuration-discovery-receiver-plugin/src/main/resources/META-INF/services/org.apache.skywalking.oap.server.library.module.ModuleDefine
@@ -16,4 +16,4 @@
 #
 #
 
-org.apache.skywalking.oap.server.recevier.configuration.discovery.ConfigurationDiscoveryModule
+org.apache.skywalking.oap.server.receiver.configuration.discovery.ConfigurationDiscoveryModule
