+++ oap-server/server-receiver-plugin/pom.xml
@@ -41,7 +41,7 @@
         <module>otel-receiver-plugin</module>
         <module>skywalking-meter-receiver-plugin</module>
         <module>skywalking-browser-receiver-plugin</module>
-        <module>skywalking-log-recevier-plugin</module>
+        <module>skywalking-log-receiver-plugin</module>
         <module>configuration-discovery-receiver-plugin</module>
         <module>skywalking-event-receiver-plugin</module>
         <module>skywalking-zabbix-receiver-plugin</module>
