+++ oap-server/server-receiver-plugin/skywalking-log-receiver-plugin/src/main/resources/META-INF/services/org.apache.skywalking.oap.server.library.module.ModuleDefine
@@ -16,4 +16,4 @@
 #
 #
 
-org.apache.skywalking.oap.server.receiver.configuration.discovery.ConfigurationDiscoveryProvider
+org.apache.skywalking.oap.server.receiver.log.module.LogModule
